Add handling for .ico files and improve error messages
Modified main.c to add support for .ico files and affecting how error messages are displayed to users. If a "get" request does not contain a resource, a descriptive message is now printed and a successful status message is sent. Improvements were also made to the handling of files not found and internal errors.