Added binary file handling and MIME type in HTTP response
A new function 'IsBinary' has been introduced to check the file type based on the extension. This function is replacing the repetitive strcmp calls for checking binary file types. 'get_file_mime' method is being used to get the MIME type of a file based on its extension for the HTTP response. This ensures accurate content-type in the HTTP header.